Add -A flag to build.sh that will attach at the end of the builds.  Also allow positional argument of "attach" to attach before builds for purposes of interactive docker usage

+++ build.sh
@@ -11,6 +11,7 @@ GPG_NAME='OnDemand Release Signing Key'
 SHOW_TASKS=false
 TASK=run
 CLEAN_OUTPUT=true
+ATTACH=false
 CLEAN_DOCKER=true
 CONTAINER="ondemand-packaging-$(whoami)"
 PACKAGES=()
@@ -45,6 +46,7 @@ function usage()
     echo "  -T         Show all tasks"
     echo "  -t TASK    Task to run, Default: $TASK"
     echo "  -C         Do not clean up output directory"
+    echo "  -A         Attach after build"
     echo "  -D         Do not clean up docker image"
     echo "  -v         Show debug information"
     echo "  -h         Show usage"
@@ -77,7 +79,7 @@ function parse_options()
 	local OPTIND=1
 	local ORIG_ARGV
 	local opt
-    while getopts "w:o:j:d:G:STt:CDvh" opt; do
+    while getopts "w:o:j:d:G:STt:CADvh" opt; do
         case "$opt" in
         w)
         	WORK_DIR="$OPTARG"
@@ -105,6 +107,9 @@ function parse_options()
             ;;
         C)
             CLEAN_OUTPUT=false
+            ;;
+        A)
+            ATTACH=true
             ;;
         D)
             CLEAN_DOCKER=false
@@ -170,7 +175,7 @@ else
     RAKE_FLAGS='-q'
 fi
 
-if $CLEAN_OUTPUT; then
+if $CLEAN_OUTPUT && [ $PACKAGES != 'attach' ]; then
     if [ -d $OUTPUT_DIR ]; then
         echo_blue "Cleaning output directory: ${OUTPUT_DIR}"
         rm -rf ${OUTPUT_DIR}/*
@@ -193,12 +198,12 @@ for p in "${PACKAGES[@]}"; do
 
     GIT_ANNEX=false
     if which git-annex 2>/dev/null 1>/dev/null ; then
-        for f in `git-annex find --include='*' ${p}`; do
+        for f in `git-annex find --include='*' ${p} 2>/dev/null`; do
             GIT_ANNEX=true
             break
         done
     fi
-    if $GIT_ANNEX ; then
+    if $GIT_ANNEX; then
         echo_blue "Unlocking git-annex sources at ${p}"
         git-annex get $p 1>/dev/null
         git-annex unlock $p 1>/dev/null
@@ -229,6 +234,11 @@ for p in "${PACKAGES[@]}"; do
         $BUILDBOX_IMAGE \
         /usr/sbin/init)
         echo_green "Container started with ID ${CONTAINER_ID}"
+    fi
+
+    if [ "$p" == 'attach' ]; then
+        DISTRIBUTIONS=''
+        ATTACH=true
     fi
 
     if [ "$TASK" == 'build:passenger_nginx' ]; then
@@ -267,6 +277,22 @@ for p in "${PACKAGES[@]}"; do
     if $GIT_ANNEX; then
         echo_blue "Locking git-annex sources at ${p}"
         git-annex lock --force $p 1>/dev/null
+    fi
+    if $ATTACH ; then
+        docker exec \
+        $TTY_ARGS \
+        -e "DISTRO=${distro}" \
+        -e "PACKAGE=${p}" \
+        -e "GPG_SIGN=${GPG_SIGN}" \
+        -e "GPG_NAME=${GPG_NAME}" \
+        -e "OOD_UID=`/usr/bin/id -u`" \
+        -e "OOD_GID=`/usr/bin/id -g`" \
+        -e "DEBUG=${DEBUG}" \
+        -e "LC_CTYPE=en_US.UTF-8" \
+        $CONTAINER \
+        /ondemand-packaging/build/inituidgid.sh \
+        /ondemand-packaging/build/setuser ood \
+        /bin/bash
     fi
     if $CLEAN_DOCKER ; then
         kill_container $CONTAINER
